+++ rust/operator-binary/src/command.rs
@@ -1,23 +1,17 @@
-use stackable_operator::commons::s3::S3ConnectionSpec;
-use stackable_operator::commons::tls::{CaCert, Tls, TlsServerVerification, TlsVerification};
-use stackable_trino_crd::authentication::TrinoAuthenticationConfig;
+use crate::catalog::config::CatalogConfig;
 use stackable_trino_crd::{
-    TrinoCluster, CONFIG_DIR_NAME, DATA_DIR_NAME, ENV_S3_ACCESS_KEY, ENV_S3_SECRET_KEY,
-    HIVE_PROPERTIES, PASSWORD_DB, RW_CONFIG_DIR_NAME, S3_SECRET_DIR_NAME, SECRET_KEY_S3_ACCESS_KEY,
-    SECRET_KEY_S3_SECRET_KEY, STACKABLE_CLIENT_TLS_DIR, STACKABLE_INTERNAL_TLS_DIR,
-    STACKABLE_MOUNT_CLIENT_TLS_DIR, STACKABLE_MOUNT_INTERNAL_TLS_DIR,
-    STACKABLE_MOUNT_SERVER_TLS_DIR, STACKABLE_SERVER_TLS_DIR, STACKABLE_TLS_STORE_PASSWORD,
+    authentication::TrinoAuthenticationConfig, TrinoCluster, CONFIG_DIR_NAME, DATA_DIR_NAME,
+    PASSWORD_DB, RW_CONFIG_DIR_NAME, STACKABLE_CLIENT_TLS_DIR, STACKABLE_INTERNAL_TLS_DIR,
+    STACKABLE_MOUNT_INTERNAL_TLS_DIR, STACKABLE_MOUNT_SERVER_TLS_DIR, STACKABLE_SERVER_TLS_DIR,
+    STACKABLE_TLS_STORE_PASSWORD, SYSTEM_TRUST_STORE, SYSTEM_TRUST_STORE_PASSWORD,
     USER_PASSWORD_DATA_DIR_NAME,
 };
 
-const STACKABLE_CLIENT_CA_CERT: &str = "stackable-client-ca-cert";
-const STACKABLE_SERVER_CA_CERT: &str = "stackable-server-ca-cert";
-const STACKABLE_INTERNAL_CA_CERT: &str = "stackable-internal-ca-cert";
+pub const STACKABLE_CLIENT_CA_CERT: &str = "stackable-client-ca-cert";
+pub const STACKABLE_SERVER_CA_CERT: &str = "stackable-server-ca-cert";
+pub const STACKABLE_INTERNAL_CA_CERT: &str = "stackable-internal-ca-cert";
 
-pub fn container_prepare_args(
-    trino: &TrinoCluster,
-    s3_spec: Option<&S3ConnectionSpec>,
-) -> Vec<String> {
+pub fn container_prepare_args(trino: &TrinoCluster, catalogs: &[CatalogConfig]) -> Vec<String> {
     let mut args = vec![];
 
     // User password data
@@ -42,29 +36,27 @@ pub fn container_prepare_args(
             STACKABLE_INTERNAL_CA_CERT,
         ));
         args.extend(chown_and_chmod(STACKABLE_INTERNAL_TLS_DIR));
-        // add cert to internal truststore
+        // Add cert to internal truststore
         if trino.tls_enabled() {
-            args.push(format!("keytool -importcert -file {STACKABLE_MOUNT_SERVER_TLS_DIR}/ca.crt -alias stackable-ca-cert -keystore {STACKABLE_INTERNAL_TLS_DIR}/truststore.p12 -storepass {STACKABLE_TLS_STORE_PASSWORD} -noprompt"));
+            args.extend(add_cert_to_stackable_truststore(
+                format!("{STACKABLE_MOUNT_SERVER_TLS_DIR}/ca.crt").as_str(),
+                STACKABLE_INTERNAL_TLS_DIR,
+                STACKABLE_CLIENT_CA_CERT,
+            ));
         }
     }
 
-    // Load S3 CA to client truststore if S3 TLS enabled
-    if let Some(s3) = s3_spec {
-        if let Some(Tls {
-            verification:
-                TlsVerification::Server(TlsServerVerification {
-                    ca_cert: CaCert::SecretClass(_),
-                }),
-        }) = &s3.tls
-        {
-            args.extend(create_key_and_trust_store(
-                STACKABLE_MOUNT_CLIENT_TLS_DIR,
-                STACKABLE_CLIENT_TLS_DIR,
-                STACKABLE_CLIENT_CA_CERT,
-            ));
-            args.extend(chown_and_chmod(STACKABLE_CLIENT_TLS_DIR));
-        }
-    }
+    // Create truststore that will be used when talking to external tools like S3
+    // It will be populated from the system truststore so that connections against public services like AWS S3 are still possible
+    args.extend(create_truststore_from_system_truststore(
+        STACKABLE_CLIENT_TLS_DIR,
+    ));
+    args.extend(chown_and_chmod(STACKABLE_CLIENT_TLS_DIR));
+
+    // Add the commands that are needed to set up the catalogs
+    catalogs.iter().for_each(|catalog| {
+        args.extend_from_slice(&catalog.init_container_extra_start_commands);
+    });
 
     args.extend(chown_and_chmod(RW_CONFIG_DIR_NAME));
     args.extend(chown_and_chmod(DATA_DIR_NAME));
@@ -73,9 +65,8 @@ pub fn container_prepare_args(
 }
 
 pub fn container_trino_args(
-    trino: &TrinoCluster,
     user_authentication: Option<&TrinoAuthenticationConfig>,
-    s3_connection_spec: Option<&S3ConnectionSpec>,
+    catalogs: &[CatalogConfig],
 ) -> Vec<String> {
     let mut args = vec![
         // copy config files to a writeable empty folder
@@ -101,40 +92,12 @@ pub fn container_trino_args(
         ));
     }
 
-    // We need to read the provided s3 credentials from the secret operator / secret class folder
-    // and export it to the required env variables in order for trino to pick them up
-    // out of the config via e.g. ${ENV:S3_ACCESS_KEY}.
-    if let Some(S3ConnectionSpec {
-        credentials: Some(_),
-        ..
-    }) = s3_connection_spec
-    {
-        args.extend(vec![
-            format!(
-                "export {env_var}=$(cat {secret_dir}/{file_name})",
-                env_var = ENV_S3_ACCESS_KEY,
-                secret_dir = S3_SECRET_DIR_NAME,
-                file_name = SECRET_KEY_S3_ACCESS_KEY
-            ),
-            format!(
-                "export {env_var}=$(cat {secret_dir}/{file_name})",
-                env_var = ENV_S3_SECRET_KEY,
-                secret_dir = S3_SECRET_DIR_NAME,
-                file_name = SECRET_KEY_S3_SECRET_KEY
-            ),
-        ]);
-    }
-
-    // hive required?
-    if trino.spec.hive_config_map_name.is_some() {
-        args.extend(vec![
-            format!( "echo Writing HIVE connect string \"hive.metastore.uri=${{HIVE}}\" to {rw_conf}/catalog/{hive_properties}",
-                     rw_conf = RW_CONFIG_DIR_NAME, hive_properties = HIVE_PROPERTIES
-            ),
-            format!( "echo \"hive.metastore.uri=${{HIVE}}\" >> {rw_conf}/catalog/{hive_properties}",
-                     rw_conf = RW_CONFIG_DIR_NAME, hive_properties = HIVE_PROPERTIES
-            )])
-    }
+    // Add the commands that are needed to set up the catalogs
+    catalogs.iter().for_each(|catalog| {
+        for (env_name, file) in &catalog.load_env_from_files {
+            args.push(format!("export {env_name}=$(cat {file})"));
+        }
+    });
 
     // start command
     args.push(format!(
@@ -148,7 +111,7 @@ pub fn container_trino_args(
 
 /// Generates the shell script to create key and truststores from the certificates provided
 /// by the secret operator.
-fn create_key_and_trust_store(
+pub fn create_key_and_trust_store(
     cert_directory: &str,
     stackable_cert_directory: &str,
     alias_name: &str,
@@ -162,6 +125,24 @@ fn create_key_and_trust_store(
         format!("cat {cert_directory}/ca.crt {cert_directory}/tls.crt > {stackable_cert_directory}/chain.crt"),
         format!("echo [{stackable_cert_directory}] Creating keystore"),
         format!("openssl pkcs12 -export -in {stackable_cert_directory}/chain.crt -inkey {cert_directory}/tls.key -out {stackable_cert_directory}/keystore.p12 --passout pass:{STACKABLE_TLS_STORE_PASSWORD}")
+    ]
+}
+
+pub fn create_truststore_from_system_truststore(truststore_directory: &str) -> Vec<String> {
+    vec![
+        format!("echo [{truststore_directory}] Creating truststore {truststore_directory}/truststore.p12 from system truststore {SYSTEM_TRUST_STORE}"),
+        format!("keytool -importkeystore -srckeystore {SYSTEM_TRUST_STORE} -srcstoretype jks -srcstorepass {SYSTEM_TRUST_STORE_PASSWORD} -destkeystore {truststore_directory}/truststore.p12 -deststoretype pkcs12 -deststorepass {STACKABLE_TLS_STORE_PASSWORD} -noprompt"),
+    ]
+}
+
+pub fn add_cert_to_stackable_truststore(
+    cert_file: &str,
+    truststore_directory: &str,
+    alias_name: &str,
+) -> Vec<String> {
+    vec![
+        format!("echo [{truststore_directory}] Adding cert from {cert_file} to truststore {truststore_directory}/truststore.p12"),
+        format!("keytool -importcert -file {cert_file} -keystore {truststore_directory}/truststore.p12 -storetype pkcs12 -noprompt -alias {alias_name} -storepass {STACKABLE_TLS_STORE_PASSWORD}"),
     ]
 }