+++ CHANGELOG.md
@@ -4,6 +4,8 @@ All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
 
+## [0.5.0] - 2022-09-07
+
 ### Added
 
 - Add support for connecting to HDFS ([#263]).